refactor: extract HasIpAttributes trait from IpOnly and CredentialsAndIp
Both authentication methods had identical getAllowedSipIps(),
setAllowedSipIps(), getTechPrefix(), setTechPrefix() implementations.
Extract into a shared trait to eliminate duplication.
